requirements.txt
Flask==3.0.2
gunicorn==21.2.0
python-dotenv==1.0.1
slack-sdk==3.27.1
python-json-logger==2.0.7
pytest==8.3.5
pytest-cov==6.0.0 